+++ css-inline/src/html/mod.rs
@@ -9,5 +9,4 @@ mod serializer;
 
 pub(crate) use self::selectors::Specificity;
 pub(crate) use document::Document;
-pub use document::DEFAULT_HTML_TREE_CAPACITY;
 pub(crate) use node::NodeId;