+++ bindings/wasm/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.10.0] - 2023-06-16
+
 ### Added
 
 - `keep_link_tags` configuration option.
@@ -137,7 +139,8 @@
 
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.9.0...HEAD
+[Unreleased]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.10.0...HEAD
+[0.10.0]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.9.0...wasm-v0.10.0
 [0.9.0]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.8.4...wasm-v0.9.0
 [0.8.4]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.8.3...wasm-v0.8.4
 [0.8.3]: https://github.com/Stranger6667/css-inline/compare/wasm-v0.8.2...wasm-v0.8.3